Create .htaccess
Are you running Apache James? 

I thought your website was down when I just tried "profanity.im", which I often do with websites because I’m lazy.

Anyway, if you do, if you add this to the root, this will redirect users to www if the have just tried without.